% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.OEFPIL.R
\name{print.OEFPIL}
\alias{print.OEFPIL}
\title{Print function for an object of class 'OEFPIL'}
\usage{
\method{print}{OEFPIL}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"OEFPIL"} (a result of a call to \code{\link{OEFPIL}}).}

\item{...}{other arguments.}
}
\value{
Function prints short summary of \code{"OEFPIL"} object into the console. In case of assigning value into the variable, it returns object of class \code{"OEFPIL"}, which is a list with components defined in \link{OEFPIL}.
}
\description{
Function prints the information about an object of class \code{"OEFPIL"}.
}
\examples{
\dontshow{
utils::example("coef.OEFPIL",echo=FALSE)}
##-- Continuing the coef.OEFPIL(.) example:

##Use of print function
print(st1)

}
\seealso{
\code{\link{OEFPIL}}
}
