% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.OEFPIL.R
\name{vcov.OEFPIL}
\alias{vcov.OEFPIL}
\title{Covariance matrix from an OEFPIL object}
\usage{
\method{vcov}{OEFPIL}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"OEFPIL"} (a result of a call to \code{\link{OEFPIL}}).}

\item{...}{other arguments.}
}
\value{
A matrix of the estimated covariances between the parameter estimates from an \code{"OEFPIL"} object.
}
\description{
Function for extracting the estimated covariance matrix from an object of class \code{"OEFPIL"}.
}
\examples{
\dontshow{
utils::example("coef.OEFPIL",echo=FALSE)}
##-- Continuing the coef.OEFPIL(.) example:

##Use of vcov function
vcov(st1)

}
\seealso{
\code{\link{OEFPIL}}
}
