% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.OEFPIL.R
\name{coef.OEFPIL}
\alias{coef.OEFPIL}
\title{Extract model coefficients from an object of class 'OEFPIL'}
\usage{
## S3 method for class 'OEFPIL'
   coef(object)
}
\arguments{
\item{object}{object of class 'OEFPIL'}
}
\value{
Vector of estimated model coefficients from on object \code{OEFPIL}
}
\description{
Function which extracts the estimated model coefficients from 'OEFPIL' object.
}
\examples{
library(MASS)

##Creating a data file
steamdata <- steam
colnames(steamdata) <- c("x","y")
startsteam <- list(b1 = 5, b2 = 8, b3 = 200)
k <- nrow(steamdata)
CM <- diag(rep(0.1,2*k))

##Creating OEFPIL object
st1 <- OEFPIL(steamdata, y ~ b1 * 10^(b2 * x/ (b3 + x)), startsteam, CM, useNLS = F)

##Use of coef function
coef(st1)

}
\seealso{
\code{\link{OEFPIL}}
}
