% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.OEFPIL.R
\name{confint.OEFPIL}
\alias{confint.OEFPIL}
\title{Calculate confidence intervals for an object of class 'OEFPIL'}
\usage{
## S3 method for class 'OEFPIL'
   confint(object, signif.level = output.form$contents$signif.level)
}
\arguments{
\item{signif.level}{numerical value or vector of significance levels for confidence intervals.}

\item{object}{object of class 'OEFPIL'.}
}
\value{
Matrix of estimated confidence intervals for model coefficients from on object \code{OEFPIL}
}
\description{
Function calculate confidence intervals for parameters counted by OEFPIL function.
}
\details{
We can add one numerical value or vector of numerical values of significance levels for confidence intervals. The default case for parameter \code{signif.level} is value from \code{OEFPIL} object.
}
\examples{
library(MASS)

##Creating a data file
steamdata <- steam
colnames(steamdata) <- c("x","y")
startsteam <- list(b1 = 5, b2 = 8, b3 = 200)
k <- nrow(steamdata)
CM <- diag(rep(0.1,2*k))

##Creating OEFPIL object
st1 <- OEFPIL(steamdata, y ~ b1 * 10^(b2 * x/ (b3 + x)), startsteam, CM, useNLS = F)

##Use of confint function
#one numerical value
confint(st1)

#vector of numerical values
confint(st1,signif.level = c(0.01,0.05,0.1))

}
\seealso{
\code{\link{OEFPIL}}
}
