% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.OEFPIL.R
\name{confint.OEFPIL}
\alias{confint.OEFPIL}
\title{Confidence intervals for OEFPIL parameters}
\usage{
## S3 method for class 'OEFPIL'
   confint(object, signif.level = output.form$contents$signif.level)
}
\arguments{
\item{signif.level}{a numerical value or a vector of significance levels for confidence intervals. If missing, a value from the input \code{OEFPIL} object is used.}

\item{object}{an object of class 'OEFPIL'.}
}
\value{
A matrix of estimated confidence intervals for model coefficients from an \code{"OEFPIL"} object. The matrix contains lower and upper confidence limits (columns) for each parameter (rows).
}
\description{
Function computes confidence intervals for the parameters counted by \code{OEFPIL} function.
}
\details{
The confidence intervals are computing under normality assumption.
}
\examples{
\dontshow{
utils::example("coef.OEFPIL",echo=FALSE)}
##-- Continuing the coef.OEFPIL(.) example:

##Use of confint function
#one numerical value
confint(st1)

#vector of numerical values
confint(st1, signif.level = c(0.01,0.05,0.1))

}
\seealso{
\code{\link{OEFPIL}}
}
