% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvplot.OEFPIL.R
\name{curvplot.OEFPIL}
\alias{curvplot.OEFPIL}
\title{Ggplot function of estimated curve of object of class 'OEFPIL'}
\usage{
curvplot.OEFPIL(object, signif.level, xx)
}
\arguments{
\item{signif.level}{numeric value or vector of significance levels for confidence interval}

\item{object}{object of class 'OEFPIL'}

\item{print}{print out result summaries in the console (default \code{TRUE})}
}
\description{
Function for plotting estimated curve with pointwise confidence bands for 'OEFPIL' object.
}
\details{
If the \code{signif.level} parameter is missing, the estimated curve will be plotted without confidence bands.
         The rest of parameters can be added similarly as in ggplot.
}
\examples{
library(MASS)
library(ggplot2)

##Creating a data file
steamdata <- steam
colnames(steamdata) <- c("x","y")
n <- nrow(steamdata)
CM1 <- diag(rep(10,2*n))
CM2 <- diag(c(rep(12,n), rep(14,n)))

##Creating OEFPIL objects
st1 <- OEFPIL(steamdata, y ~ b1 * 10^(b2 * x/ (b3 + x)), list(b1 = 5, b2 = 8, b3 = 200),
             CM1, useNLS = F)
st2 <- OEFPIL(steamdata, y ~ b1 * 10^(b2 * x/ (b3 + x)), list(b1 = 5, b2 = 8, b3 = 200),
             CM2, useNLS = F)

##Use of curvplot.OEFPIL function on object of class 'OEFPIL'
curvplot.OEFPIL(st1, signif.level = 0.05)

##Use of curvplot.OEFPIL function on object of class 'OEFPIL' with different arguments
curvplot.OEFPIL(st2, signif.level = c(0.01,0.05), xx = seq(0,110,1))

##Use of curvplot.OEFPIL function with additional arguments as for ggplot2
curvplot.OEFPIL(st1, signif.level = 0.05) + labs(x = "New x label") + labs(title = "New graph title")

}
\seealso{
\code{\link{OEFPIL}}, \code{\link{paramplot.OEFPIL}} and code{\link{plot.OEFPIL}}.
}
