% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoIndent.OEFPIL.R
\name{NanoIndent.OEFPIL}
\alias{NanoIndent.OEFPIL}
\title{Calculation of pestimation of parameters of unloaded curve during Nanoindentation}
\usage{
NanoIndent.OEFPIL(data, alpha.start, m.start, hp.start, unload.data = F,ucut = 0.98,
                         lcut = 0.4, CM, uh = 0.5, uF = 0.001, max.iter = 100, see.iter.val = F,
                         save.file.name, th = .Machine$double.eps ^ (2 / 3), signif.level = 0.05,
                         useNLS = T)
}
\arguments{
\item{data}{Data file can be any object of type \code{data.frame} with 2 columns or \code{list} with 2 elements.}

\item{unload.data}{logical value (default FALSE ), which we set up to F in the case if we have both loaded and unloaded curve. In case of TRUE, we have only unloaded curve.}

\item{ucut}{numerical value, indicating the upper bound of cut off.}

\item{lcut}{numerical value, indicating the lower bound of cut off.}

\item{CM}{covariance matrix (does not change in iteration process).}

\item{uh}{standard deviation of depth}

\item{uF}{standard deviation of force}

\item{max.iter}{maximum number of iterations}

\item{see.iter.val}{logical value (default \code{TRUE}) indicating if we want to display and save partial results of algorithm.}

\item{save.file.name}{name of the file for saving results.}

\item{th}{numerical value, indicating threshold necessary for estimation stoppage.}

\item{signif.level}{significance level for confidence interval}

\item{useNLS}{logical. If \code{TRUE} (the default value), function will set up starting parameters calculated by \code{\link{nlsLM}} function (nonlinear least square estimation).}

\item{form}{an object of class \code{\link{formula}} (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{name.start}{starting values of estimating parameters.}
}
\value{
Returns an object of class \code{'OEPFIL'} is a list containing at least the following components
\itemize{
   \item name_est estimations of model parameters
   \item name_upgraded.start.val estimations of model parameters
   \item it_num number of iterations
   \item CI_parameters list of confidence intervals for estimated parameters (significance level is based on parameter signif.level)
   \item logs warnings or messages of events, which happen during the run of our function
   \item \code{name_previous} values from the previous iterative step
}
In addition we get \code{contents}, which is list of outputs as original values and entered parameters, which are usable in other estimation process.
If we set up parameter \code{useNLS} to TRUE, the start.values which enters in the estimation process will be claculated by \code{\link{nlsLM}} function. Otherwise the \code{start.values} and \code{name_upgraded.start.val} will be the same.
In the case, if we add estimated parameters with NaN, NA, Inf or -Inf, we end up with error message.
}
\description{
Function calculates parameters of unloaded curve during nanoindentation by iterated linearization.
}
\details{
Function prepared data (clean and cut them off) for \code{\link{OEPFIL}}, for parameter estimation.
 Functional dependence of parameters is fixed in
the form: \code{f = alpha * (h - hp) ^ m}, where \code{f} is strength and \code{h} depth
(on the unloaded part of curve). User has an option to enter his own starting values of
parameters, or the values will be set up by algorithm.

In case, we do not add our own covariance matrix, it will be calculated by the algorithm with use of \code{u_h} and \code{u_f}.
}
\examples{
## We use "uncut" data file: "silicaberk.txt" which is part of OEFPIL package
## Preparing parameter for OEFPIL function
unload.data = T
ucut = 0.98
lcut = 0.2
uh = 0.5
uF = 0.001
max.iter = 100
see.iter.val = F
th = 0.001
signif.level = 0.05
useNLS = T


##use of NanoIndent function with default parameters
output.form.NI <- NanoIndent.OEFPIL(data)

##use of summary function for clear output
summary(output.form.NI)


##use of NanoIndent function with our predefined parameters
output.form.NI <- NanoIndent.OEFPIL(silicaBerk, unload.data = unload.data, ucut = ucut,
                            lcut = lcut, uh = uh, uF = uF, max.iter = max.iter,
                            see.iter.val = see.iter.val, th = th,
                            signif.level = signif.level, useNLS = useNLS)

##use of summary function for clear output
summary(output.form.NI)

##plot of estimated functions
plot(output.form.NI, signif.level = signif.level)

}
\references{
Kubáček, L. and Kubáčková, L. (2000) \emph{Statistika a metrologie}, Univerzita Palackého v Olomouci.

   Köning1, R., Wimmer, G. and Witkovský, V. (2014) \emph{Ellipse fitting by nonlinear constraints to demodulate quadrature homodyne interferometer signals and to determine the statistical uncertainty of the interferometric phase}, Measurement Science and Technology.
}
\seealso{
\code{\link{OEFPIL}}
}
