% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortresiduals.OEFPIL.R
\name{ortresiduals.OEFPIL}
\alias{ortresiduals.OEFPIL}
\title{Calculation of orthogonal residuals from an object of class 'OEFPIL'}
\usage{
ortresiduals.OEFPIL(object, a)
}
\arguments{
\item{a}{numeric value, for defining minimization interval for the \code{\link{optimize}} function (if not defined, default value 0.05 * range(x) is used). Must be positive}

\item{object}{object of class 'OEFPIL'}
}
\value{
Returns an object of type list containing at least the following components
 \itemize{\item \code{x.ores} numerical vector of x coordinates of points, where the minimal distance is realized
           \item  \code{o.resid} numerical vector of orthogonal residuals (minimal Euclidean distances between data points and estimated function)
           \item \code{SSort} orthogonal sum of squares.
           }
}
\description{
Function for calculating orthogonal residuals of an 'OEFPIL' object.
}
\examples{
library(MASS)

##Creating a data file
steamdata <- steam
colnames(steamdata) <- c("x","y")
n <- nrow(steamdata)
CM1 <- diag(rep(0.1,2*n))

##Creating OEFPIL object
st1 <- OEFPIL(steamdata, y ~ b1 * 10^(b2 * x/ (b3 + x)), list(b1 = 5, b2 = 8, b3 = 200),
             CM1, useNLS = F)

##Use ortresiduals.OEFPIL function on the OEFPIL object, with specified value 'a'
ortresiduals.OEFPIL(st1,5)

##Use ortresiduals.OEFPIL function without value 'a' (defaut.value = 0.05 * range(x))
ortresiduals.OEFPIL(st1)

##Choice of too narrow interval. Result can be misleading!
ortresiduals.OEFPIL(st1,0.5)
}
\seealso{
\code{\link{OEFPIL}}
}
