% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramplot.OEFPIL.R
\name{paramplot.OEFPIL}
\alias{paramplot.OEFPIL}
\title{Parameter estimation graph with ggplot of object of class 'OEFPIL'}
\usage{
paramplot.OEFPIL(object)
}
\arguments{
\item{object}{object or \code{list} of class 'OEFPIL'}
}
\description{
Function for parameter estimation graph with error bars (plus minus standard deviation) with ggplot.
}
\details{
In the case, if we add estimated parameters with NaN, NA, Inf or -Inf, we get empty plot.
}
\examples{
library(MASS)
steamdata <- steam
colnames(steamdata) <- c("x","y")
n <- nrow(steamdata)
CM1 <- diag(rep(0.1,2*n))
CM2 <- diag(c(rep(0.2^2,n), rep(0.1^2,n)))
st1 <- OEFPIL(steamdata, y ~ b1 * 10^(b2 * x/ (b3 + x)), list(b1 = 5, b2 = 8, b3 = 200),
             CM1, useNLS = F)
st2 <- OEFPIL(steamdata, y ~ b1 * 10^(b2 * x/ (b3 + x)), list(b1 = 5, b2 = 8, b3 = 200),
             CM2, useNLS = F)

##Use of paramplot.OEFPIL function on object of class 'OEFPIL'
paramplot.OEFPIL(st2)

##Use of paramplot.OEFPIL function on list of object of class 'OEFPIL'
paramplot.OEFPIL(list(st1,st2))

##Now we look at the case when we plot parameters with Nan values (OEPFIL function does not converge).
startsteam <- list(b1 = 0.1, b2 = 5, b3 = 200)
st3 <- OEFPIL(steamdata, y ~ b1 * 10^(b2 * x/ (b3 + x)), list(b1 = 0.1, b2 = 5, b3 = 200),
              CM1, useNLS = F)
paramplot.OEFPIL(st3)

}
\seealso{
\code{\link{OEFPIL}}, \code{\link{curvplot.OEFPIL}} and code{\link{plot.OEFPIL}}.
}
