% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.OEFPIL.R
\name{plot.OEFPIL}
\alias{plot.OEFPIL}
\title{Plot the estimate from an OEPFIL object}
\usage{
## S3 method for class 'OEFPIL'
    plot(object, xx, signif.level,...)
}
\arguments{
\item{xx}{a sequence of x-coordinates of points for computing and plotting confidence bands. If missing, the default sequence \code{seq(from = min(x), to = max(x), length.out = 301)} is used.}

\item{signif.level}{a numerical value or a vector of significance levels for confidence bands.}

\item{...}{additional arguments (same as in \link{plot} function) affecting the plot.}

\item{object}{an object of class \code{"OEFPIL"} (a result of a call to \code{\link{OEFPIL}}).}
}
\value{
Returns an object of type list containing at least the following components
 \itemize{ \item \code{xx} a numerical vector of points where bands are calculated.
           \item \code{yy} a numerical vector with values of estimated function in \code{xx}.
           \item \code{PointwiseCB} a matrix of pointwise confidence bands at points \code{xx}.
           }
}
\description{
Plot of the iterated linearization estimate of a function from an \code{"OEFPIL"} object with pointwise confidence bands.
}
\details{
If the \code{signif.level} argument is missing, the value is set to 0.05, but the confidence bands are not plotted.
         The confidence bands are computing under normality assumption.
}
\examples{
\dontshow{
utils::example("coef.OEFPIL",echo=FALSE)}
##-- Continuing the coef.OEFPIL(.) example:

##Use of plot function with default parameters, signif.level is not set up...No confidence bands are plotted
plot(st1)

##Use of plot function with different parameters
plot(st1, seq(0,113,0.1), signif.level = c(0.01,0.05), main = "Graph of estimated function")

##Return values of plot function
(a <- plot(st1, signif.level = 0.05))


}
\seealso{
\code{\link{OEFPIL}}
}
