% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.OEFPIL.R
\name{plot.OEFPIL}
\alias{plot.OEFPIL}
\title{Plot of the iterated linearization estimate from an OEPFIL object}
\usage{
## S3 method for class 'OEFPIL'
    plot(object, xx, signif.level,...)
}
\arguments{
\item{xx}{numerical vector of points, where confidence bands will be calculated and plotted.}

\item{signif.level}{numerical value or vector of significance levels for confidence intervals.}

\item{...}{additional arguments (same as \link{plot} function) affecting the plot.}

\item{object}{object of class 'OEFPIL'}
}
\value{
Returns an object of type list containing at least the following components
 \itemize{ \item  \code{xx} points where bands are calculated.
           \item \code{yy} values of estimated function.
           \item \code{PointwiseCB} matrix of pointwise confidence bands at points \code{xx}.
           }
}
\description{
Plot of the iterated linearization estimate from an OEPFIL object.
}
\details{
If the signif.level parameter is missing, we do not plot any confidence bands, although the values will be set to 0.05.
}
\examples{
library(MASS)

##Creating a data file
steamdata <- steam
colnames(steamdata) <- c("x","y")
startsteam <- list(b1 = 5, b2 = 8, b3 = 200)
k <- nrow(steamdata)
CM <- diag(rep(12,2*k))

##Creating OEFPIL object
st1 <- OEFPIL(steamdata, y ~ b1 * 10^(b2 * x/ (b3 + x)), startsteam, CM, useNLS = F)

##Use of plot function with default parameters, signif.level is not set up...No confidence bands
plot(st1)

##Use of plot function with different parameters
plot(st1, seq(0,113,0.1), signif.level = c(0.01,0.05), main = "Graph of estimated function")

##Return values of plot function
(a <- plot(st1, signif.level = 0.05))


}
\seealso{
\code{\link{OEFPIL}}
}
