% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.OEFPIL.R
\name{print.OEFPIL}
\alias{print.OEFPIL}
\title{Print function for an object of class 'OEFPIL'}
\usage{
\method{print}{OEFPIL}(object)
}
\arguments{
\item{object}{an object of class \code{"OEFPIL"} (a result of a call to \code{\link{OEFPIL}}).}
}
\description{
Function prints the information about an object of class \code{"OEFPIL"}.
   print(object)
}
\details{
Function prints the formula and estimated parameters of the model from an \code{OEFPIL} object.
}
\examples{
\dontshow{
utils::example("coef.OEFPIL",echo=FALSE)}
##-- Continuing the coef.OEFPIL(.) example:

##Use of print function
print(st1)

}
\seealso{
\code{\link{OEFPIL}}
}
