% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.OEFPIL.R
\name{print.OEFPIL}
\alias{print.OEFPIL}
\title{Print function for an object of class 'OEFPIL'}
\usage{
## S3 method for class 'OEFPIL'
   print(object)
}
\arguments{
\item{object}{object of class 'OEFPIL'.}
}
\description{
Function will print out the information about object of class 'OEFPIL'.
}
\details{
Function will print the formula and estimated parameters of model from an object \code{OEFPIL}.
}
\examples{
\dontshow{
utils::example("coef.OEFPIL",echo=FALSE)}
##-- Continuing the coef.OEFPIL(.) example:

##Use of print function
print(st1)

}
\seealso{
\code{\link{OEFPIL}}
}
