% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silica2098-data.R
\docType{data}
\name{silica2098}
\alias{silica2098}
\title{Nanoindentation measurements data}
\format{
The data frame contains two columns:
\describe{
  \item{Depth}{nanoindentation depth, in nanometers}
  \item{Fload}{load, in milinewtons}
}
}
\source{
Czech Metrology Institute, Brno, Czech Republic.
}
\usage{
silica2098
}
\description{
Load and depth measured with Berkovich diamond tip on fused silica material.
 Data contains only cutted part of unloading indentation curve (20 - 98\%, as recommended in
 ISO 14577 standard). The relationship between \emph{Fload} and \emph{Depth} \emph{h} is given by equation
 \eqn{Fload = \alpha (h - h_p)^{m}}, where  \emph{h_p} is permanent indentation depth after
 removal of the load, \eqn{\alpha} is fitting constant related to the indenter geometry and
 power law exponent \emph{m} should be from (1,2) interval.
}
\examples{
attach(silica2098)
plot(Depth, Fload, main = 'Graph of nanoindentation data', xlab = 'Depth (nm)', ylab = 'Load (mN)',
col = 'darkgreen', cex = 1)
}
\references{
ISO/IEC: \emph{14577-1:2015 Metallic materials – Instrumented indentation test for hardness and materials parameters – Part 1: Test method} (ISO/IEC, Internation Organisation for Standardisation, 2015).
}
\keyword{datasets}
