% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.OEFPIL.R
\name{summary.OEFPIL}
\alias{summary.OEFPIL}
\title{Summary function from an object of class 'OEFPIL'}
\usage{
## S3 method for class 'OEFPIL'
   summary(object, signif.level = output.form$contents$signif.level, print = T)
}
\arguments{
\item{signif.level}{significance level for confidence interval}

\item{print}{print out result summaries in the console (default \code{TRUE})}

\item{object}{object of class 'OEFPIL'}
}
\value{
Returns an object of type list containing at least the following components
 \itemize{\item \code{param_Est} is the (numerical) vector of estimated model parameters
           \item  \code{sd} standard deviations for estimated model parameters.
           \item \code{cov.m_Est} covariance matrix of estimated model parameters.
           \item \code{it_num} number of iterations
           \item \code{CI_parameters} matrix of lower and upper bounds for confidence intervals.
           }
}
\description{
Function for fast and clean output of all basic information of 'OEFPIL' object.
}
\details{
The default case for parameter \code{signif.level} is value from \code{OEFPIL} object.
}
\examples{
library(MASS)

##Creating a data file
steamdata <- steam
colnames(steamdata) <- c("x","y")
startsteam <- list(b1 = 5, b2 = 8, b3 = 200)
k <- nrow(steamdata)
CM <- diag(rep(0.1,2*k))

##Creating OEFPIL object
st1 <- OEFPIL(steamdata, y ~ b1 * 10^(b2 * x/ (b3 + x)), startsteam, CM, useNLS = F)

##Use of summary function with default parameters
summary(st1)

##Use of summary function with different parameters
summary(st1, signif.level = 0.05, print = F)

}
\seealso{
\code{\link{OEFPIL}}
}
