% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.OEFPIL.R
\name{vcov.OEFPIL}
\alias{vcov.OEFPIL}
\title{Extract covariance matrix from an object of class 'OEFPIL'}
\usage{
## S3 method for class 'OEFPIL'
   vcov(object)
}
\arguments{
\item{object}{object of class 'OEFPIL'}
}
\description{
Function for extracting the estimated covariance matrix from 'OEFPIL' object.
}
\examples{
\dontshow{
utils::example("coef.OEFPIL",echo=FALSE)}
##-- Continuing the coef.OEFPIL(.) example:

##Use of vcov function
vcov(st1)

}
\seealso{
\code{\link{OEFPIL}}
}
